Moved class, instance, and ipl_library to value.rs
